+++ local/provision/docker_provision.sh
@@ -4,6 +4,8 @@ set -e
 
 ## FOR LAMP STACK
 # docker-compose exec web bash /var/www/html/docker/provision/bin/sudo_web_tasks.sh
+# echo "Set password for user www-data"
+# docker-compose exec web /bin/bash -c "passwd www-data"
 # docker-compose exec web bash /var/www/html/docker/provision/bin/disable_xdebug.sh
 # docker-compose exec mysql bash /devbox/provision/bin/db_tasks.sh
 # docker-compose exec -u $UID web bash /var/www/html/docker/provision/bin/web_tasks.sh